fix(slack): repair Enterprise Grid CI checks
