+++ extensions/slack/src/monitor/provider.auth-test-token.test.ts
@@ -342,7 +342,7 @@ describe("auth.test boot call", () => {
       appToken: "xapp-1-A1-opaque",
     });
     await vi.waitFor(() => expect(getSlackTestState().appStartMock).toHaveBeenCalledTimes(1));
-    expect([...getSlackTestState().interactionRegistrations].sort()).toEqual([
+    expect([...getSlackTestState().interactionRegistrations].toSorted()).toEqual([
       "action",
       "command",
       "shortcut",
@@ -979,117 +979,6 @@ describe("connected identity health", () => {
       }),
     );
     expect(getSlackHandlers().has("reaction_added")).toBe(true);
-    await stopSlackMonitor(monitor);
-  });
-
-  it("does not adopt Enterprise identity from Bolt event context", async () => {
-    resetSlackTestState({
-      channels: {
-        slack: {
-          mode: "http",
-          signingSecret: "test-signing-secret",
-          dmPolicy: "disabled",
-          groupPolicy: "open",
-          channels: { C12345678: { allow: true, requireMention: true } },
-        },
-      },
-    });
-    const client = getSlackClient();
-    client.auth.test.mockRejectedValue(new Error("request_timeout"));
-    client.conversations.info.mockResolvedValueOnce({
-      channel: { name: "general", is_channel: true },
-    });
-    const { replyMock, sendMock } = getSlackTestState();
-    replyMock.mockResolvedValue({ text: "unexpected" });
-    const setStatus = vi.fn();
-    const monitor = startSlackMonitor(monitorSlackProvider, { setStatus });
-    const handler = await getSlackHandlerOrThrow("message");
-
-    await handler({
-      event: {
-        type: "message",
-        user: "UOTHER123",
-        text: "<@UCONTEXT> status",
-        ts: "100.000",
-        channel: "C12345678",
-        channel_type: "channel",
-      },
-      context: {
-        botUserId: "UCONTEXT",
-        botId: "BCONTEXT",
-        isEnterpriseInstall: true,
-        enterpriseId: "E_ENTERPRISE",
-        teamId: "TWORKSPACE",
-      },
-      body: { api_app_id: "A_ENTERPRISE" },
-      client,
-    });
-
-    expect(setStatus).not.toHaveBeenCalledWith(expect.objectContaining({ lifecycle: "ready" }));
-    expect(replyMock).not.toHaveBeenCalled();
-    expect(sendMock).not.toHaveBeenCalled();
-    await stopSlackMonitor(monitor);
-  });
-
-  it("adopts Bolt identity from the first HTTP event and restores mention detection", async () => {
-    resetSlackTestState({
-      channels: {
-        slack: {
-          mode: "http",
-          signingSecret: "test-signing-secret",
-          groupPolicy: "open",
-          requireMention: true,
-        },
-      },
-    });
-    const client = getSlackClient();
-    client.auth.test.mockRejectedValue(new Error("request_timeout"));
-    client.conversations.info.mockResolvedValueOnce({
-      channel: { name: "general", is_channel: true },
-    });
-    const { replyMock, sendMock } = getSlackTestState();
-    replyMock.mockResolvedValue({ text: "identity restored" });
-    const setStatus = vi.fn();
-    const monitor = startSlackMonitor(monitorSlackProvider, { setStatus });
-    const handler = await getSlackHandlerOrThrow("message");
-
-    expect(setStatus).toHaveBeenCalledWith({
-      connected: true,
-      lastConnectedAt: expect.any(Number),
-      terminalDisconnect: true,
-      lifecycle: "blocked",
-      lastError: "request_timeout",
-    });
-    expect(setStatus).not.toHaveBeenCalledWith(expect.objectContaining({ connected: false }));
-
-    await handler({
-      event: {
-        type: "message",
-        user: "U_OTHER",
-        text: "<@URECOVERED> status",
-        ts: "999999.123",
-        channel: "C12345678",
-        channel_type: "channel",
-      },
-      context: {
-        botUserId: "URECOVERED",
-        botId: "BRECOVERED",
-        teamId: "T12345678",
-        isEnterpriseInstall: false,
-      },
-      body: {},
-    });
-
-    expect(setStatus).toHaveBeenCalledWith({
-      running: true,
-      connected: true,
-      lastConnectedAt: expect.any(Number),
-      terminalDisconnect: undefined,
-      lifecycle: "ready",
-      lastError: null,
-    });
-    expect(getSlackHandlers().has("reaction_added")).toBe(true);
-    await vi.waitFor(() => expect(sendMock).toHaveBeenCalledTimes(1));
     await stopSlackMonitor(monitor);
   });
 });
